backend/pyproject.toml
[project]
name = "f1-recruiting-backend"
version = "0.1.0"
description = "F1 Driver Recruiting Platform Backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.0",
    # Validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "email-validator>=2.2.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.12",
    # LLM / RAG
    "openai>=1.50.0",
    "sentence-transformers>=3.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    # Background tasks
    "celery[redis]>=5.4.0",
    "redis>=5.1.0",
    # Utilities
    "structlog>=24.4.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
